backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
